requirements.txt
asgiref==3.6.0
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.1
idna==3.4
iniconfig==1.1.1
numpy==1.23.4
opencv-python==4.6.0.66
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
pytz==2022.7
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.12
